refactor: organizes ColorThumb exports in their respective folder
