+++ src/components/ui/color-thumb/index.ts
@@ -1,0 +1,2 @@
+export { default as ColorThumb } from './ColorThumb.vue';
+export type { ColorThumbProps } from './types';
