+++ src/components/ui/color-thumb/types.ts
@@ -1,0 +1,9 @@
+import type { Size } from '../../../types';
+
+export interface ColorThumbProps {
+  hexadecimal?: string;
+  hexadecimalSecondary?: string;
+  image?: string;
+  size?: Size;
+  width?: number | string;
+}
